+++ src/DiscoToolbarBundle.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-/* #############################################################################
+/* **********************************************************************
+ * 
+ *     █▀▀▄  ▀                     ▀▀█▀▀           ▀█  █
+ *     █  █ ▀█  ▄▀▀▄ ▄▀▀▄ ▄▀▀▄       █   ▄▀▀▄ ▄▀▀▄  █  █▀▀▄ ▄▀▀▄ █▄▀
+ *     █  █  █   ▀▄  █    █  █       █   █  █ █  █  █  █  █  ▄▄█ █
+ *     █▄▄▀ ▄█▄ ▀▄▄▀ ▀▄▄▀ ▀▄▄▀       █   ▀▄▄▀ ▀▄▄▀ ▄█▄ █▄▄▀ ▀▄▄▀ █
  *
- *   █▀▀▄  ▀                     █▀▀▄           █▀▀▄
- *   █  █ ▀█  ▄▀▀▄ ▄▀▀▄ ▄▀▀▄     █  █ ▄▀▀▄ █  █ █▀▀▄ ▄▀▀▄ █▄▀
- *   █  █  █   ▀▄  █    █  █     █  █ █▀▀  █  █ █  █  ▄▄█ █
- *   █▄▄▀ ▄█▄ ▀▄▄▀ ▀▄▄▀ ▀▄▄▀     █▄▄▀ ▀▄▄▀ ▀▄▀  █▄▄▀ ▀▄▄▀ █
- *
- *     Customizable developer toolbar for Symfony projects
+ * *******  Customizable developer toolbar for Symfony projects  ********
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2025 Marcin Orlowski
+ * @copyright 2025-2026 Marcin Orlowski
  * @license   https://opensource.org/license/mit MIT
- * @link      https://github.com/MarcinOrlowski/php-discotoolbar-symfony
+ * @link      https://github.com/MarcinOrlowski/php-discobar-symfony
  *
- * ########################################################################## */
+ * ******************************************************************* */
 
 namespace MarcinOrlowski\DiscoToolbar;