+++ src/Listener/ExceptionListener.php
@@ -2,6 +2,22 @@
 
 declare(strict_types=1);
 
+/* **********************************************************************
+ * 
+ *     █▀▀▄  ▀                     ▀▀█▀▀           ▀█  █
+ *     █  █ ▀█  ▄▀▀▄ ▄▀▀▄ ▄▀▀▄       █   ▄▀▀▄ ▄▀▀▄  █  █▀▀▄ ▄▀▀▄ █▄▀
+ *     █  █  █   ▀▄  █    █  █       █   █  █ █  █  █  █  █  ▄▄█ █
+ *     █▄▄▀ ▄█▄ ▀▄▄▀ ▀▄▄▀ ▀▄▄▀       █   ▀▄▄▀ ▀▄▄▀ ▄█▄ █▄▄▀ ▀▄▄▀ █
+ *
+ * *******  Customizable developer toolbar for Symfony projects  ********
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @copyright 2025-2026 Marcin Orlowski
+ * @license   https://opensource.org/license/mit MIT
+ * @link      https://github.com/MarcinOrlowski/php-discobar-symfony
+ *
+ * ******************************************************************* */
+
 namespace MarcinOrlowski\DiscoToolbar\Listener;
 
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;