+++ CHANGES.md
@@ -1,10 +1,12 @@
+![img](img/banner.webp)
+
 # Changelog
 
 ## 2.0.0 (TBD)
 
+- [#044] Renamed package to `marcin-orlowski/disco-toolbar-symfony`
 - [#047] Rebranded from DiscoDevBar to DiscoToolbar
 - [#041] Added Symfony 8 support
-- [#044] Renamed package to `marcin-orlowski/disco-toolbar-symfony`
 - [#042] Added project logo!
 - [#035] Added customizable background colors for breathing stripes
 - [#033] Added close widget type to dismiss toolbar